Relocate classes in home feature module
