+++ feature/home/src/main/java/akio/apps/myrun/feature/home/userhome/di/UserHomeFeatureComponent.kt
@@ -1,4 +1,4 @@
-package akio.apps.myrun.feature.home.wiring
+package akio.apps.myrun.feature.home.userhome.di
 
 import akio.apps.myrun.data.activity.ActivityDataModule
 import akio.apps.myrun.data.authentication.AuthenticationDataModule
